Task_1/requirements.txt
python-dotenv==1.0.1
google-generativeai==0.4.1
supabase==2.0.4
requests>=2.31.0
sqlparse==0.4.4 
